+++ src/main/java/br/edu/fib/bibliotecajavamvc/service/LivroService.java
@@ -1,5 +1,6 @@
 package br.edu.fib.bibliotecajavamvc.service;
 
+import br.edu.fib.bibliotecajavamvc.excecao.LivroPossuiAssociacoesException;
 import br.edu.fib.bibliotecajavamvc.model.Livro;
 import br.edu.fib.bibliotecajavamvc.repository.LivroRepository;
 import br.edu.fib.bibliotecajavamvc.storage.FotoStorage;
@@ -24,14 +25,16 @@ public class LivroService {
 
     public List<Livro> pesquisarLivrosSemEmprestimoVigente() {
         List<Livro> livros = livroRepository.listaLivrosSemEmprestimoVigente();
-        List<Livro> livrosComMediaDeAvaliacoes = livroRepository.listaMediaAvaliacoesPorLivro();
+        preencheMediaDasAvaliacoes(livros);
+        return livros;
+    }
 
+    private void preencheMediaDasAvaliacoes(List<Livro> livros) {
+        List<Livro> livrosComMediaDeAvaliacoes = livroRepository.listaMediaAvaliacoesPorLivro();
         livros.forEach(livro -> livro.setMediaAvaliacoes(
                 livrosComMediaDeAvaliacoes.stream().filter(livroAvaliacao -> livroAvaliacao.getId().equals(livro.getId()))
                         .findFirst().orElse(new Livro()).getMediaAvaliacoes())
         );
-
-        return livros;
     }
 
     public void salvar(Livro livro) {
@@ -46,7 +49,11 @@ public class LivroService {
         }
     }
 
-    public void excluir(Long id) {
+    public void excluir(Long id) throws LivroPossuiAssociacoesException {
+        if (livroRepository.livroSemAssociacoes(id) == null) {
+            throw new LivroPossuiAssociacoesException();
+        }
+
         this.livroRepository.delete(id);
     }
